refactor: combine masked dep with enemy dep
