+++ lc-hax/Scripts/Patches/Dependencies/EnemyDependencyPatch.cs
@@ -5,7 +5,6 @@ using Hax;
 
 [HarmonyPatch(typeof(EnemyAI))]
 class EnemyDependencyPatch {
-
     [HarmonyPatch(nameof(EnemyAI.OnDestroy))]
     static void Prefix(EnemyAI __instance) {
         _ = Helper.Enemies.Remove(__instance);
@@ -13,6 +12,19 @@ class EnemyDependencyPatch {
 
     [HarmonyPatch(nameof(EnemyAI.Start))]
     static void Postfix(EnemyAI __instance) {
+        _ = Helper.Enemies.Add(__instance);
+    }
+}
+
+[HarmonyPatch(typeof(MaskedPlayerEnemy))]
+class MaskedDependencyPatch {
+    [HarmonyPatch(nameof(MaskedPlayerEnemy.OnDestroy))]
+    static void Prefix(MaskedPlayerEnemy __instance) {
+        _ = Helper.Enemies.Remove(__instance);
+    }
+
+    [HarmonyPatch(nameof(MaskedPlayerEnemy.Start))]
+    static void Postfix(MaskedPlayerEnemy __instance) {
         _ = Helper.Enemies.Add(__instance);
     }
 }
